refactor: ActionType, EntityType ê 경로이동
